Allow checked binding handler to handle grouped elements
If the DOM selector for an input matches multiple elements, eg input[name="gender"], then the checked binding handler would be provided an array of elements but would only operate on the first element within the array. In the case of setting the value, we only need to set the value of the matched element. For getting the value, we can rely on the target of the event to specify with which element was interacted.